backend/requirements.txt
# Core dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==1.10.12
python-dotenv==1.0.0
starlette==0.27.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0
asyncpg==0.28.0

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==41.0.3
email-validator==2.0.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2
aiohttp==3.8.5
tenacity==8.2.3

# LLM integrations
tiktoken==0.5.1
transformers==4.32.1
torch==2.0.1
sentencepiece==0.1.99

# Caching and rate limiting
redis==5.0.0
aioredis==2.0.1

# File handling
python-multipart==0.0.6
boto3==1.28.41

# Logging
loguru==0.7.0

# Testing
pytest==7.4.1
pytest-asyncio==0.21.1
httpx==0.24.1
pytest-cov==4.1.0

# Development
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1